move DiscordBot and logger setup out of main
