+++ .gitignore
@@ -8,3 +8,4 @@
 perf_*.txt
 dc_token.txt
 client_id.txt
+latest.log
